% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{apply_transform}
\alias{apply_transform}
\title{Transform second argument by first.}
\usage{
apply_transform(vps, dframe, ..., parallelCluster = NULL)
}
\arguments{
\item{vps}{vtreat pipe step, object defining transform.}

\item{dframe}{data.frame, data to transform}

\item{...}{not used, forces later arguments to bind by name.}

\item{parallelCluster}{optional, parallel cluster to run on.}
}
\value{
transformed dframe
}
\description{
Apply first argument to second as a transform.
}
