% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{catScoreFrame}
\alias{catScoreFrame}
\title{return pseudo R-squareds
Can use a parallel cluster (and sends only sections of data to cluster elements)}
\usage{
catScoreFrame(df, namevec, yC, yTarget, weights = c(),
  parallelCluster = c())
}
\arguments{
\item{df}{numeric data frame (no NAs/NULLs) effective variables}

\item{namevec}{names of columns to process}

\item{yC}{(no NAs/NULLs) outcome values}

\item{yTarget}{scalar target for yC to match (yC==tTarget is goal)}

\item{weights}{(optional) numeric, non-negative, no NAs/NULLs at least two positive positions}

\item{parallelCluster}{(optional) a cluster object created by package parallel or package snow}
}
\value{
pseud R-squareds catScpre(y,weights), it is an R-squared so 1 is good 0 is bad
}
\description{
return pseudo R-squareds
Can use a parallel cluster (and sends only sections of data to cluster elements)
}
\seealso{
\code{\link{hold1OutMeans}}
}

