% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{designTreatmentsN}
\alias{designTreatmentsN}
\title{designTreatmentsN}
\usage{
designTreatmentsN(dframe, varlist, outcomename, weights = c(),
  minFraction = 0.02, smFactor = 0, maxMissing = 0.04, collarProb = 0,
  scoreVars = TRUE, maxScoreSize = 1000000L, verbose = TRUE)
}
\arguments{
\item{dframe}{Data frame to learn treatments from (training data).}

\item{varlist}{Names of columns to treat (effective variables).}

\item{outcomename}{Name of column holding outcome variable.}

\item{weights}{optional training weights for each row}

\item{minFraction}{optional minimum frequency a categorical level must have to be converted to an indicator column.}

\item{smFactor}{optional smoothing factor for impact coding models.}

\item{maxMissing}{optional maximum fraction (by data weight) of a categorical variable that are allowed before switching from indicators to impact coding.}

\item{collarProb}{what fraction of the data (pseudo-probability) to collar data at (<0.5).}

\item{scoreVars}{optional if TRUE attempt to estimate individual variable utility.}

\item{maxScoreSize}{optional maximum size for treated variable scoring frame}

\item{verbose}{if TRUE print progress.}
}
\value{
treatment plan (for use with prepare)
}
\description{
Function to design variable treatments for binary prediction of a
numeric outcome.  Data frame is assumed to have only atomic columns
except for dates (which are converted to numeric).
Note: each column is processed independently of all others.
}
\examples{
dTrainN <- data.frame(x=c('a','a','a','a','b','b','b'),
    z=c(1,2,3,4,5,6,7),y=c(0,0,0,1,0,1,1))
dTestN <- data.frame(x=c('a','b','c',NA),
    z=c(10,20,30,NA))
treatmentsN = designTreatmentsN(dTrainN,colnames(dTrainN),'y')
dTrainNTreated <- prepare(treatmentsN,dTrainN)
dTestNTreated <- prepare(treatmentsN,dTestN)
}
\seealso{
\code{\link{prepare}} \code{\link{designTreatmentsC}}
}

