% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.wmean}
\alias{.wmean}
\title{Compute weighted mean}
\usage{
.wmean(x, weights = NULL)
}
\arguments{
\item{x}{numeric vector without NA to compute mean of}

\item{weights}{weights vector (or NULL)}
}
\value{
weighted mean
}
\description{
Compute the weighted mean of x.
}
\examples{

.wmean(c(1, 2, 3))

}
\keyword{internal}
