% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{fit}
\alias{fit}
\title{Fit first arguemnt to data in second argument.}
\usage{
fit(vps, dframe, ..., weights = NULL, parallelCluster = NULL)
}
\arguments{
\item{vps}{vtreat pipe step, mutable object to fit to.}

\item{dframe}{data.frame, data to fit from.}

\item{...}{not used, forces later arguments to bind by name.}

\item{weights}{optional, per-dframe data weights.}

\item{parallelCluster}{optional, parallel cluster to run on.}
}
\value{
vps
}
\description{
Update the state of first argument to have learned or fit from second argument.
}
