% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{fit_transform}
\alias{fit_transform}
\title{Fit and transform in a cross-validated manner.}
\usage{
fit_transform(vps, dframe, ..., weights = NULL, parallelCluster = NULL)
}
\arguments{
\item{vps}{vtreat pipe step, object specifying fit.}

\item{dframe}{data.frame, data to fit from.}

\item{...}{not used, forces later arguments to bind by name.}

\item{weights}{optional, per-dframe data weights.}

\item{parallelCluster}{optional, parallel cluster to run on.}
}
\value{
@return named list containing: vtreat_pipe_step and cross_frame
}
\description{
Update the state of first argument to have learned or fit from second argument, and compute a cross
validated example of such a transform.
}
\details{
Note: input vps is not altered, fit is in returned list.
}
