% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquery_treatment.R
\name{flatten_fn_list}
\alias{flatten_fn_list}
\title{Flatten a list of functions onto d.}
\usage{
flatten_fn_list(d, fnlist)
}
\arguments{
\item{d}{object (usually a data souce)}

\item{fnlist}{a list of functions}
}
\value{
fnlist[[length(fnlist)]](flatten_fn_list(d, fnlist[[-length(fnlist)]]) (or d if length(fnlist)<1)
}
\description{
Flatten a list of functions onto d.
}
\seealso{
\code{\link{as_rquery_plan}}
}
\keyword{internal}
