% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{format.vtreatment}
\alias{format.vtreatment}
\title{Display treatment plan.}
\usage{
\method{format}{vtreatment}(x, ...)
}
\arguments{
\item{...}{additional args (to match general signature).}

\item{vtreat}{treatment plan}
}
\description{
Display treatment plan.
}

