% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdHocTreatments.R
\name{gamXform}
\alias{gamXform}
\title{Build a one variable GAM transform}
\usage{
gamXform(x, y, family = stats::gaussian())
}
\arguments{
\item{x}{numeric vector input}

\item{y}{numeric vector (same lenght as x) target output}

\item{family}{mgcv/stats family object specifying the distribution and link to use in fitting}
}
\value{
increasing function mapping x to y
}
\description{
Depends on mgcv and stats
}
\examples{

d <- data.frame(x=0.05*(1:100))
d$y <- sin(d$x)
m <- gamXform(d$x,d$y)
d$pred <- m(d$x)
# ggplot(data=d,mapping=aes(x=x)) + geom_point(aes(y=y)) +
#   geom_line(aes(y=pred),color='blue')

d$yCat <- d$y>0.1
mC <- gamXform(d$x, d$yC, family= stats::binomial(link='logit'))
d$predC <- mC(d$x)
# ggplot(data=d,mapping=aes(x=predC,color=yCat)) + geom_density()

}

