% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{get_feature_names}
\alias{get_feature_names}
\title{Return feasible feature names.}
\usage{
get_feature_names(vps)
}
\arguments{
\item{vps}{vtreat pipe step, mutable object to read from.}
}
\value{
feature names
}
\description{
Return previously fit feature names.
}
