% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{get_score_frame.vtreat_pipe_step}
\alias{get_score_frame.vtreat_pipe_step}
\title{Return score frame from vps.}
\usage{
get_score_frame.vtreat_pipe_step(vps)
}
\arguments{
\item{vps}{vtreat pipe step, mutable object to read from.}
}
\value{
score frame
}
\description{
Return previously fit score frame.
}
