% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{get_transform}
\alias{get_transform}
\title{Return underlying transform from vps.}
\usage{
get_transform(vps)
}
\arguments{
\item{vps}{vtreat pipe step, mutable object to read from.}
}
\value{
transform
}
\description{
Return previously fit transform.
}
