% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{getSplitPlanAppLabels}
\alias{getSplitPlanAppLabels}
\title{read application labels off a split plan.}
\usage{
getSplitPlanAppLabels(nRow, plan)
}
\arguments{
\item{nRow}{number of rows in original data.frame.}

\item{plan}{split plan}
}
\value{
vector of labels

plan <- oneWayHoldout(3,NULL,NULL,NULL)
getSplitPlanAppLabels(3,plan)
}
\description{
read application labels off a split plan.
}

