% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{hold1OutMeans}
\alias{hold1OutMeans}
\title{Return a vector of length(y) where the i-th entry is the weighted mean
of all but the i-th y.  Usefull for normalizing PRESS style statistics.}
\usage{
hold1OutMeans(y, weights)
}
\arguments{
\item{y}{values to average (should not have NAs).}

\item{weights}{data weighing (should not have NAs, be non-negative and not all zero).}
}
\description{
Return a vector of length(y) where the i-th entry is the weighted mean
of all but the i-th y.  Usefull for normalizing PRESS style statistics.
}

