% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{kWayCrossValidation}
\alias{kWayCrossValidation}
\title{k-fold cross validation, a splitFunction in the sense of vtreat::buildEvalSets}
\usage{
kWayCrossValidation(nSplits, nRows, dframe, y)
}
\arguments{
\item{nSplits}{number of groups to split into.}

\item{nRows}{number of rows to split.}

\item{dframe}{original data frame.}

\item{y}{numeric outcome variable.}
}
\value{
split plan
}
\description{
k-fold cross validation, a splitFunction in the sense of vtreat::buildEvalSets
}

