% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{kWayStratifiedY}
\alias{kWayStratifiedY}
\title{k-fold cross validation stratified on y, a splitFunction in the sense of vtreat::buildEvalSets}
\usage{
kWayStratifiedY(nRows, nSplits, dframe, y)
}
\arguments{
\item{nRows}{number of rows to split.}

\item{nSplits}{number of groups to split into.}

\item{dframe}{original data frame (ignored).}

\item{y}{numeric outcome variable try to have equidistributed in each split.}
}
\value{
split plan
}
\description{
k-fold cross validation stratified on y, a splitFunction in the sense of vtreat::buildEvalSets
}
\examples{

set.seed(23255)
d <- data.frame(y=sin(1:100))
pStrat <- kWayStratifiedY(nrow(d),5,d,d$y)
getAppLabels <- function(nRow,plan) {
  labels <- numeric(nRow)
  for(i in seq_len(length(plan))) {
    labels[plan[[i]]$app] <- i
  }
  labels
}
d$stratGroup <- getAppLabels(nrow(d),pStrat)
pSimple <- kWayCrossValidation(nrow(d),5,d,d$y)
d$simpleGroup <- getAppLabels(nrow(d),pSimple)
summary(tapply(d$y,d$simpleGroup,mean))
# ggplot(data=d,aes(x=y,color=as.factor(simpleGroup))) + 
#   geom_density() + ggtitle('simple grouping')
summary(tapply(d$y,d$stratGroup,mean))
# ggplot(data=d,aes(x=y,color=as.factor(stratGroup))) + 
#   geom_density() + ggtitle('y-stratified grouping')

}

