% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linScore}
\alias{linScore}
\title{Return in-sample linear stats and scaling.}
\usage{
linScore(varName, xcol, ycol, weights)
}
\arguments{
\item{varName}{name of variable}

\item{xcol}{numeric vector of inputs (no NA/NULL/NaN)}

\item{ycol}{numeric vector of outcomes (no NA/NULL/NaN)}

\item{weights}{numeric vector of data weights (no NA/NULL/NaN, all>0.0)}
}
\description{
Return in-sample linear stats and scaling.
}

