% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customCoder.R
\name{makeCustomCoderNum}
\alias{makeCustomCoderNum}
\title{Make a numeric input custom coder.}
\usage{
makeCustomCoderNum(
  ...,
  customCode,
  coder,
  codeSeq,
  v,
  vcolin,
  zoY,
  zC,
  zTarget,
  weights = NULL,
  catScaling = FALSE
)
}
\arguments{
\item{...}{not used, force arguments to be set by name}

\item{customCode}{code name}

\item{coder}{user supplied variable re-coder (see vignette for type signature)}

\item{codeSeq}{argments to custom coder}

\item{v}{variable name}

\item{vcolin}{data column, numeric}

\item{zoY}{outcome column as numeric}

\item{zC}{if classification outcome column as character}

\item{zTarget}{if classification target class}

\item{weights}{per-row weights}

\item{catScaling}{optional, if TRUE use glm() linkspace, if FALSE use lm() for scaling.}
}
\value{
wrapped custom coder
}
\description{
Make a numeric input custom coder.
}
