% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{makekWayCrossValidationGroupedByColumn}
\alias{makekWayCrossValidationGroupedByColumn}
\title{k-fold cross validation, respecting (never splitting) groupingColumn.}
\usage{
makekWayCrossValidationGroupedByColumn(groupingColumnName)
}
\arguments{
\item{groupingColumnName}{name of column to group by.}

\item{nRows}{number of rows to split (>1).}

\item{nSplits}{number of groups to split into (>1,<=nRows).}
}
\value{
split plan
}
\description{
k-fold cross validation, respecting (never splitting) groupingColumn.
}
\examples{

d <- data.frame(y=sin(1:100))
d$group <- floor(seq_len(nrow(d))/5)
splitter <- makekWayCrossValidationGroupedByColumn('group')
split <- splitter(nrow(d),5,d,d$y)
d$splitLabel <- vtreat::getSplitPlanAppLabels(nrow(d),split)
rowSums(table(d$group,d$splitLabel)>0)

}

