% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquery_treatment.R
\name{materialize_treated}
\alias{materialize_treated}
\title{Materialize a treated data frame remotely.}
\usage{
materialize_treated(db, rqplan, data_source, result_table_name, ...,
  temporary = FALSE, overwrite = TRUE)
}
\arguments{
\item{db}{a db handle.}

\item{rqplan}{an query plan}

\item{data_source}{relop, data source (usually a relop_table_source).}

\item{result_table_name}{character, table name to land result in}

\item{...}{force later arguments to bind by name.}

\item{temporary}{logical, if TRUE try to make result temporary.}

\item{overwrite}{logical, if TRUE try to overwrite result.}
}
\value{
description of treated table.
}
\description{
Materialize a treated data frame remotely.
}
\seealso{
\code{\link{as_rquery}}
}
