% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdHocTreatments.R
\name{monotoneXform}
\alias{monotoneXform}
\title{Build a monotone transform}
\usage{
monotoneXform(x, y, k = NULL)
}
\arguments{
\item{x}{numeric vector input}

\item{y}{numeric vector (same lenght as x) target output}

\item{k}{integer if not null approximate number of knots to prune to}
}
\value{
monotone function mapping x to y
}
\description{
Depends on stats
}
\examples{

d <- data.frame(x=c(15,1,-12,NA,5,NaN))
d$y <- -log(d$x)
m <- monotoneXform(d$x,d$y)
d2 <- data.frame(x=-20:20)
d2$pred <- m(d2$x)
# ggplot() + geom_point(data=d,mapping=aes(x=x,y=y)) + 
#  geom_line(data=d2,mapping=aes(x=x,y=pred),color='blue')

}

