% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdHocTreatments.R
\name{monotoneXformD}
\alias{monotoneXformD}
\title{Build a monotone transform transform in a chosen direction}
\usage{
monotoneXformD(x, y, decreasing = FALSE, k = NULL)
}
\arguments{
\item{x}{numeric vector input}

\item{y}{numeric vector (same lenght as x) target output}

\item{decreasing}{logical if TRUE build decreasing instead}

\item{k}{integer if not null approximate number of knots to prune to}
}
\value{
monotone function mapping x to y
}
\description{
Depends on stats
}
\examples{

d <- data.frame(x=c(3,1,-12,NA,5,NaN))
d$y <- log(d$x)
m1 <- monotoneXformD(d$x,d$y)
m2 <- monotoneXformD(d$x,d$y,decreasing=TRUE)
d$pred1 <- m1(d$x)
d$pred2 <- m2(d$x)

}

