% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{multinomial_parameters}
\alias{multinomial_parameters}
\title{vtreat multinomial parameters.}
\usage{
multinomial_parameters(user_params = NULL)
}
\arguments{
\item{user_params}{list of user overrides.}
}
\value{
filled out parameter list
}
\description{
A list of settings and values for vtreat multinomial classification fitting. 
Please see
\url{https://github.com/WinVector/vtreat/blob/main/Examples/fit_transform/fit_transform_api.md},
\code{\link{mkCrossFrameMExperiment}} and
\code{\link{prepare.multinomial_plan}} for details.
}
