% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_columns_into_frame.R
\name{patch_columns_into_frame}
\alias{patch_columns_into_frame}
\title{Patch columns into data.frame.}
\usage{
patch_columns_into_frame(orig_frame, new_frame)
}
\arguments{
\item{orig_frame}{data.frame to patch into.}

\item{new_frame}{data.frame to take replacement columns from.}
}
\value{
patched data.frame
}
\description{
Add columns from new_frame into old_frame, replacing any
columns with matching names in orig_frame with values from
new_frame.
}
\examples{

orig_frame <- data.frame(x = 1, y = 2)
new_frame <- data.frame(y = 3, z = 4)
patch_columns_into_frame(orig_frame, new_frame)

}
