% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_in_frame.R
\name{patch_in_frame}
\alias{patch_in_frame}
\title{Patch new columns into a data frame.}
\usage{
patch_in_frame(orig_frame, new_frame)
}
\arguments{
\item{orig_frame}{data.frame}

\item{new_frame}{data.frame}
}
\value{
merged data.frame
}
\description{
Add columns from new_frame to orig_frame, replacing
any columns with matching names.
}
\examples{

patch_in_frame(data.frame(x = 1, y = 2), data.frame(x = 3, z = 4))

}
