% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_pooling.R
\name{ppCoderC}
\alias{ppCoderC}
\title{Solve a categorical partial pooling problem.}
\usage{
ppCoderC(v, vcol, y, w = NULL)
}
\arguments{
\item{v}{character variable name}

\item{vcol}{character, independent or input variable}

\item{y}{logical, dependent or outcome variable to predict}

\item{w}{row/example weights}
}
\value{
scored training data column
}
\description{
Please see \url{http://www.win-vector.com/blog/2017/09/custom-level-coding-in-vtreat/} and
\url{http://www.win-vector.com/blog/2017/09/partial-pooling-for-lower-variance-variable-encoding/}.
}
