% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_pooling.R
\name{ppCoderN}
\alias{ppCoderN}
\title{Solve a numeric partial pooling problem.}
\usage{
ppCoderN(v, vcol, y, w = NULL)
}
\arguments{
\item{v}{character variable name}

\item{vcol}{character, independent or input variable}

\item{y}{numeric, dependent or outcome variable to predict}

\item{w}{row/example weights}
}
\value{
scored training data column
}
\description{
Please see \url{https://win-vector.com/2017/09/25/custom-level-coding-in-vtreat/} and
\url{https://win-vector.com/2017/09/28/partial-pooling-for-lower-variance-variable-encoding/}.
}
