% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{pre_comp_xval}
\alias{pre_comp_xval}
\title{Pre-computed cross-plan (so same split happens each time).}
\usage{
pre_comp_xval(nRows, nSplits, splitplan)
}
\arguments{
\item{nRows}{number of rows to split (integer >1).}

\item{nSplits}{number of groups to split into (ignored).}

\item{splitplan}{split plan to actually use}
}
\value{
splitplan
}
\description{
Pre-computed cross-plan (so same split happens each time).
}
\examples{

p1 <- oneWayHoldout(3,NULL,NULL,NULL)
p2 <- pre_comp_xval(3, 3, p1)
p2(3, 3)

}
