% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtreat.R
\name{prepare}
\alias{prepare}
\title{Apply treatments and restrict to useful variables.}
\usage{
prepare(treatmentplan, dframe, ...)
}
\arguments{
\item{treatmentplan}{Plan built by designTreantmentsC() or designTreatmentsN()}

\item{dframe}{Data frame to be treated}

\item{...}{no additional arguments, declared to forced named binding of later arguments}
}
\description{
Apply treatments and restrict to useful variables.
}
\seealso{
\code{\link{prepare.treatmentplan}}, \code{\link{prepare.simple_plan}}, \code{\link{prepare.multinomial_plan}}
}
