% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{pressStatOfBestLinearFit}
\alias{pressStatOfBestLinearFit}
\title{Compute the PRESS statistic of a 1-variable linear model}
\usage{
pressStatOfBestLinearFit(x, y, weights, normalizationStrat = "total")
}
\arguments{
\item{x}{numeric (no NAs/NULLs) effective variable}

\item{y}{numeric (no NAs/NULLs) outcome variable}

\item{weights}{numeric, non-negative, no NAs/NULLs at least two positive positions}

\item{normalizationStrat}{'none': no normalization (traditional PRESS), 'total': divide by total variation, 'holdout': divide by 1-hold out variation (PRESS-line, larger than total variation)}
}
\value{
PRESS statistic of model y ~ a*x + b divided by pressStatOfBestConstant(y,weights)
}
\description{
Compute the PRESS statistic of a 1-variable linear model
}

