% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{pressStatOfBestLinearFitFrame}
\alias{pressStatOfBestLinearFitFrame}
\title{Compute the PRESS R-squared statistic of a 1-variable linear model
Can use a parallel cluster (and sends only sections of data to cluster elements)}
\usage{
pressStatOfBestLinearFitFrame(df, namevec, y, weights = c(),
  normalizationStrat = "total", parallelCluster = c())
}
\arguments{
\item{df}{numeric data frame (no NAs/NULLs) effective variables}

\item{namevec}{names of columns to process}

\item{y}{numeric (no NAs/NULLs) outcome variable}

\item{weights}{(optional) numeric, non-negative, no NAs/NULLs at least two positive positions}

\item{normalizationStrat}{(optional) 'none': no normalization (traditional PRESS), 'total': divide by total variation, 'holdout': divide by 1-hold out variation (PRESS-line, larger than total variation)}

\item{parallelCluster}{(optional) a cluster object created by package parallel or package snow}
}
\value{
PRESS statistics of model y ~ a*x + b divided by pressStatOfBestConstant(y,weights), it is an R-squared so 1 is good 0 is bad
}
\description{
Compute the PRESS R-squared statistic of a 1-variable linear model
Can use a parallel cluster (and sends only sections of data to cluster elements)
}
\seealso{
\code{\link{hold1OutMeans}}
}

