% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{pressStatOfCategoricalVariable}
\alias{pressStatOfCategoricalVariable}
\title{Compute the PRESS statistic a single categorical model.   Tries to prevent some of the test/train leakage in scoring
(so apply this directly to a categorical variable, and don't score an impact coded variable).}
\usage{
pressStatOfCategoricalVariable(vcolin, y, weights,
  normalizationStrat = "total", smoothingTerm = 0.5)
}
\arguments{
\item{vcolin}{character}

\item{y}{numeric vectors (no NAs/NULLs)}

\item{weights}{numeric, non-negative, no NAs/NULLs at least two positive positions}

\item{normalizationStrat}{'none': no normalization (traditional PRESS), 'total': divide by total variation, 'holdout': divide by 1-hold out variation (PRESS-line, larger than total variation)}

\item{smoothingTerm}{scalar >= 0}
}
\value{
PRESS statistic of model y ~ x divided by pressStatOfBestConstant(y,weights)
}
\description{
Compute the PRESS statistic a single categorical model.   Tries to prevent some of the test/train leakage in scoring
(so apply this directly to a categorical variable, and don't score an impact coded variable).
}
\seealso{
\code{\link{hold1OutMeans}} \code{\link{pressStatOfBestLinearFit}}
}

