% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_class.R
\name{print.multinomial_plan}
\alias{print.multinomial_plan}
\title{Print treatmentplan.}
\usage{
\method{print}{multinomial_plan}(x, ...)
}
\arguments{
\item{x}{treatmentplan}

\item{...}{additional args (to match general signature).}
}
\description{
Print treatmentplan.
}
