% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_missing_Z.R
\name{print.simple_plan}
\alias{print.simple_plan}
\title{Print treatmentplan.}
\usage{
\method{print}{simple_plan}(x, ...)
}
\arguments{
\item{x}{treatmentplan}

\item{...}{additional args (to match general signature).}
}
\description{
Print treatmentplan.
}
