% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{print.vtreatment}
\alias{print.vtreatment}
\title{Print treatment plan.}
\usage{
\method{print}{vtreatment}(x, ...)
}
\arguments{
\item{x}{treatment plan}

\item{...}{additional args (to match general signature).}
}
\description{
Print treatment plan.
}

