% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{problemAppPlan}
\alias{problemAppPlan}
\title{check if appPlan is a good partition of 1:nRows into nSplits groups}
\usage{
problemAppPlan(nRows, nSplits, appPlan, strict)
}
\arguments{
\item{nRows}{number of rows to partition}

\item{nSplits}{number of sets to partition into}

\item{appPlan}{partition to critique}

\item{strict}{logical, if true expect application data to be a partition and training data to be maximal}
}
\value{
problem with partition (null if good)
}
\description{
check if appPlan is a good partition of 1:nRows into nSplits groups
}
\examples{

plan <- oneWayHoldout(3,NULL,NULL,NULL)
problemAppPlan(3,3,plan,TRUE)

}

