% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{problemWithPartition}
\alias{problemWithPartition}
\title{check if splits is a good partition of 1:nRows into ncross groups}
\usage{
problemWithPartition(nRows, ncross, splits)
}
\arguments{
\item{nRows}{number of rows to partition}

\item{ncross}{number of sets to partition into}

\item{splits}{partition to critique}
}
\value{
problem with partition (null if good)
}
\description{
check if splits is a good partition of 1:nRows into ncross groups
}

