% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{regression_parameters}
\alias{regression_parameters}
\title{vtreat regression parameters.}
\usage{
regression_parameters(user_params = NULL)
}
\arguments{
\item{user_params}{list of user overrides.}
}
\value{
filled out parameter list
}
\description{
A list of settings and values for vtreat regression fitting. 
See
\code{\link{mkCrossFrameCExperiment}}, 
\code{\link{designTreatmentsC}}, and
\code{\link{mkCrossFrameNExperiment}}, 
\code{\link{designTreatmentsN}},
\code{\link{prepare.treatmentplan}} for details.
}
