% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{show.vtreatment}
\alias{show.vtreatment}
\title{Display treatment plan.}
\usage{
show.vtreatment(vtreat, ...)
}
\arguments{
\item{vtreat}{treatment plan}

\item{...}{additional args (to match general signature).}
}
\description{
Display treatment plan.
}

