% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_variable.R
\name{spline_variable}
\alias{spline_variable}
\title{Spline numeric variable}
\usage{
spline_variable(varName, x, y, w = NULL)
}
\arguments{
\item{varName}{character, name of variable}

\item{x}{numeric input (not empty, no NAs).}

\item{y}{numeric or castable to such (same length as x no NAs), output to match}

\item{w}{numeric positive, same length as x (weights, can be NULL)}
}
\value{
spline y prediction
}
\description{
Return a spline approximation of data.
}
