% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{unsupervised_parameters}
\alias{unsupervised_parameters}
\title{vtreat unsupervisedparameters.}
\usage{
unsupervised_parameters(user_params = NULL)
}
\arguments{
\item{user_params}{list of user overrides.}
}
\value{
filled out parameter list
}
\description{
A list of settings and values for vtreat unsupervised fitting. 
See
\code{\link{designTreatmentsZ}}, and
\code{\link{prepare.treatmentplan}} for details.
}
