% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{UnsupervisedTreatment}
\alias{UnsupervisedTreatment}
\title{Stateful object for designing and applying unsupervised treatments.}
\usage{
UnsupervisedTreatment(
  ...,
  var_list,
  cols_to_copy = NULL,
  params = NULL,
  imputation_map = NULL
)
}
\arguments{
\item{...}{not used, force arguments to be specified by name.}

\item{var_list}{Names of columns to treat (effective variables).}

\item{cols_to_copy}{list of extra columns to copy.}

\item{params}{parameters list from \code{unsupervised_parameters}}

\item{imputation_map}{map from column names to functions of signature f(values: numeric, weights: numeric), simple missing value imputers.}
}
\description{
Hold settings are results for unsupervised data preparation.
}
\details{
Please see
\url{https://github.com/WinVector/vtreat/blob/master/Examples/fit_transform/fit_transform_api.md},
\code{\link{designTreatmentsZ}} and
\code{\link{prepare.treatmentplan}} for details.

Note: for \code{UnsupervisedTreatment} \code{fit_transform(d)} is implemented
as \code{fit(d)$transform(d)}.
}
