% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmented_variable.R
\name{variable_values}
\alias{variable_values}
\title{Return variable evaluations.}
\usage{
variable_values(sf)
}
\arguments{
\item{sf}{scoreFrame from from vtreat treatments}
}
\value{
per-original varaible evaluations
}
\description{
Return variable evaluations.
}
