% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{vnames}
\alias{vnames}
\title{New treated variable names.}
\usage{
vnames(x, ...)
}
\arguments{
\item{x}{vtreatment item.}

\item{...}{additional args (to match general signature).}
}
\description{
New treated variable names.
}

