% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{vorig}
\alias{vorig}
\title{Original variable name from a treatmentplan$treatment item.}
\usage{
vorig(x)
}
\arguments{
\item{x}{vtreatment item.}
}
\description{
Original variable name from a treatmentplan$treatment item.
}
\seealso{
\code{\link{designTreatmentsC}} \code{\link{designTreatmentsN}} \code{\link{getNewVarNames}}
}

