% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtreat.R
\docType{package}
\name{vtreat-package}
\alias{vtreat}
\alias{vtreat-package}
\title{vtreat: A Statistically Sound 'data.frame' Processor/Conditioner}
\description{
A 'data.frame' processor/conditioner that prepares real-world data for predictive modeling in a statistically sound manner.
'vtreat' prepares variables so that data has fewer exceptional cases, making
it easier to safely use models in production. Common problems 'vtreat' defends
against: 'Inf', 'NA', too many categorical levels, rare categorical levels, and new
categorical levels (levels seen during application, but not during training).
'vtreat::prepare' should be used as you would use 'model.matrix'.
}
\details{
For more information:
\itemize{
  \item \code{vignette('vtreat', package='vtreat')}
  \item \code{vignette(package='vtreat')}
  \item Website: \url{https://github.com/WinVector/vtreat} }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WinVector/vtreat/}
  \item \url{https://winvector.github.io/vtreat/}
  \item Report bugs at \url{https://github.com/WinVector/vtreat/issues}
}

}
\author{
\strong{Maintainer}: John Mount \email{jmount@win-vector.com}

Authors:
\itemize{
  \item Nina Zumel \email{nzumel@win-vector.com}
}

Other contributors:
\itemize{
  \item  Win-Vector LLC [copyright holder]
}

}
