% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{vtreat_parameters}
\alias{vtreat_parameters}
\title{vtreat parameters.}
\usage{
vtreat_parameters(user_params = NULL)
}
\arguments{
\item{user_params}{list of user overrides.}
}
\value{
filled out parameter list
}
\description{
A list of settings and values for vtreat supervised fitting. 
See
\code{\link{mkCrossFrameCExperiment}}, 
\code{\link{designTreatmentsC}}, and
\code{\link{prepare.treatmentplan}} for details.
}
