% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquery_treatment.R
\name{as_rquery}
\alias{as_rquery}
\title{Convert a vtreat tstep into an rquery operation tree.}
\usage{
as_rquery(tstep, d, ...)
}
\arguments{
\item{tstep}{vtreat treatment plan}

\item{d}{relop data source (usually of class relop_table_source)}

\item{...}{not used, force any later arguments to bind to names.}
}
\value{
rquery optree
}
\description{
Convert a vtreat tstep into an rquery operation tree.
}
\examples{

if(requireNamespace("rquery", quietly = TRUE)) {
   dTrainC <- data.frame(x= c('a','a','a','b',NA,'b'),
                         z= c(1,2,NA,4,5,6),
                         y= c(FALSE,FALSE,TRUE,FALSE,TRUE,TRUE))
   treatmentsC <- designTreatmentsC(dTrainC, colnames(dTrainC),'y',TRUE)
   ops <- as_rquery(treatmentsC, rquery::local_td(dTrainC))
   cat(format(ops))
   if(requireNamespace("rqdatatable", quietly = TRUE)) {
      treated <- rqdatatable::ex_data_table(ops)
      print(treated[])
   }
   if(requireNamespace("DBI", quietly = TRUE) &&
      requireNamespace("DBI", quietly = TRUE)) {
      db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
      old_o <- options(list("rquery.rquery_db_executor" = list(db = db)))
      cat(format(rquery::to_sql(ops, db)))
      treated <- rquery::execute(dTrainC, ops, allow_executor = FALSE)
      print(treated)
      options(old_o)
      DBI::dbDisconnect(db)
   }
}

}
