% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catScore}
\alias{catScore}
\title{return a pseudo R-squared from a 1 variable logistic regression}
\usage{
catScore(varName, x, yC, yTarget, weights)
}
\arguments{
\item{varName}{name of variable}

\item{x}{numeric (no NAs/NULLs) effective variable}

\item{yC}{(no NAs/NULLs) outcome variable}

\item{yTarget}{scalar target for yC to match (yC==tTarget is goal)}

\item{weights}{(optional) numeric, non-negative, no NAs/NULLs at least two positive positions}
}
\value{
cross-validated pseudo-Rsquared estimate of a 1-variable logistic regression
}
\description{
return a pseudo R-squared from a 1 variable logistic regression
}

