% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft.R
\name{classification_parameters}
\alias{classification_parameters}
\title{vtreat classification parameters.}
\usage{
classification_parameters(user_params = NULL)
}
\arguments{
\item{user_params}{list of user overrides.}
}
\value{
filled out parameter list
}
\description{
A list of settings and values for vtreat binomial classification fitting. 
Please see
\url{https://github.com/WinVector/vtreat/blob/master/Examples/fit_transform/fit_transform_api.md},
\code{\link{mkCrossFrameCExperiment}}, 
\code{\link{designTreatmentsC}}, and
\code{\link{prepare.treatmentplan}} for details.
}
