% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{designTreatmentsC}
\alias{designTreatmentsC}
\title{designTreatmentsC}
\usage{
designTreatmentsC(dframe, varlist, outcomename, outcometarget, weights = c(),
  minFraction = 0.02, smFactor = 0, maxMissing = 0.04, collarProb = 0,
  scoreVars = TRUE, maxScoreSize = 100000L, verbose = TRUE,
  parallelCluster = NULL)
}
\arguments{
\item{dframe}{Data frame to learn treatments from (training data).}

\item{varlist}{Names of columns to treat (effective variables).}

\item{outcomename}{Name of column holding outcome variable.}

\item{outcometarget}{Value/level of outcome to be considered "success"}

\item{weights}{optional training weights for each row}

\item{minFraction}{optional minimum frequency a categorical level must have to be converted to an indicator column.}

\item{smFactor}{optional smoothing factor for impact coding models.}

\item{maxMissing}{optional maximum fraction (by data weight) of a categorical variable that are allowed before switching from indicators to impact coding.}

\item{collarProb}{what fraction of the data (pseudo-probability) to collar data at (<0.5).}

\item{scoreVars}{optional if TRUE attempt to estimate individual variable utility.}

\item{maxScoreSize}{optional maximum size for treated variable scoring frame}

\item{verbose}{if TRUE print progress.}

\item{parallelCluster}{(optional) a cluster object created by package parallel or package snow}
}
\value{
treatment plan (for use with prepare)
}
\description{
Function to design variable treatments for binary prediction of a
categorical outcome.  Data frame is assumed to have only atomic columns
except for dates (which are converted to numeric).
}
\details{
The main fields are mostly vectors with names (all with the same names in the same order):

- vars : (character array without names) names of variables (in same order as names on the other diagnostic vectors)
- varMoves : logical TRUE if the variable varied during training, only variables that move will be in the treated frame
- PRESSRsquared : a PRESS-held out R-squared of a linear fit from each variable to the y-value.  Scores of zero and below are very bad, scores near one are very good.
- catPseudoRSquared : the pseudo-Rsquared (deviance ratio) of each variable in turn logisticly regressed against the categorical y target.  Similar ot the PRESSRsquared this attempts to be a hold-out statistic

See the vtreat vignette for a bit more detail and a worked example.
}
\examples{
dTrainC <- data.frame(x=c('a','a','a','b','b','b'),
   z=c(1,2,3,4,5,6),
   y=c(FALSE,FALSE,TRUE,FALSE,TRUE,TRUE))
dTestC <- data.frame(x=c('a','b','c',NA),
   z=c(10,20,30,NA))
treatmentsC <- designTreatmentsC(dTrainC,colnames(dTrainC),'y',TRUE)
dTrainCTreated <- prepare(treatmentsC,dTrainC)
dTestCTreated <- prepare(treatmentsC,dTestC)
}
\seealso{
\code{\link{prepare}} \code{\link{designTreatmentsN}}
}

