% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{designTreatmentsC}
\alias{designTreatmentsC}
\title{designTreatmentsC}
\usage{
designTreatmentsC(dframe, varlist, outcomename, outcometarget, weights = c(),
  minFraction = 0.02, smFactor = 0, rareCount = 4, maxMissing = 0.04,
  collarProb = 0, returnXFrame = FALSE, scale = FALSE, doCollar = TRUE,
  verbose = TRUE, parallelCluster = NULL)
}
\arguments{
\item{dframe}{Data frame to learn treatments from (training data).}

\item{varlist}{Names of columns to treat (effective variables).}

\item{outcomename}{Name of column holding outcome variable.}

\item{outcometarget}{Value/level of outcome to be considered "success"}

\item{weights}{optional training weights for each row}

\item{minFraction}{optional minimum frequency a categorical level must have to be converted to an indicator column.}

\item{smFactor}{optional smoothing factor for impact coding models.}

\item{rareCount}{optional integer, supress direct effects of level of this count or less.}

\item{maxMissing}{optional maximum fraction (by data weight) of a categorical variable that are allowed before switching from indicators to impact coding.}

\item{collarProb}{what fraction of the data (pseudo-probability) to collar data at (<0.5).}

\item{returnXFrame}{optional if TRUE return out of sample transformed frame.}

\item{scale}{logical optional controls scaling for scoring and returnXFrame}

\item{doCollar}{logical optional controls collaring for scoring and returnXFrame}

\item{verbose}{if TRUE print progress.}

\item{parallelCluster}{(optional) a cluster object created by package parallel or package snow}
}
\value{
treatment plan (for use with prepare)
}
\description{
Function to design variable treatments for binary prediction of a
categorical outcome.  Data frame is assumed to have only atomic columns
except for dates (which are converted to numeric).
}
\details{
The main fields are mostly vectors with names (all with the same names in the same order):

- vars : (character array without names) names of variables (in same order as names on the other diagnostic vectors)
- varMoves : logical TRUE if the variable varied during hold out scoring, only variables that move will be in the treated frame
- #' - sig : an estimate signficance of effect

See the vtreat vignette for a bit more detail and a worked example.
}
\examples{
dTrainC <- data.frame(x=c('a','a','a','b','b','b'),
   z=c(1,2,3,4,5,6),
   y=c(FALSE,FALSE,TRUE,FALSE,TRUE,TRUE))
dTestC <- data.frame(x=c('a','b','c',NA),
   z=c(10,20,30,NA))
treatmentsC <- designTreatmentsC(dTrainC,colnames(dTrainC),'y',TRUE)
dTrainCTreated <- prepare(treatmentsC,dTrainC,pruneSig=0.99)
dTestCTreated <- prepare(treatmentsC,dTestC,pruneSig=0.99)
}
\seealso{
\code{\link{prepare}} \code{\link{designTreatmentsN}}
}

