% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtreat.R
\name{designTreatmentsZ}
\alias{designTreatmentsZ}
\title{Design variable treatments with no outcome variable.}
\usage{
designTreatmentsZ(dframe, varlist, ..., minFraction = 0.02, weights = c(),
  rareCount = 0, collarProb = 0, verbose = TRUE, parallelCluster = NULL)
}
\arguments{
\item{dframe}{Data frame to learn treatments from (training data), must have at least 1 row.}

\item{varlist}{Names of columns to treat (effective variables).}

\item{...}{no additional arguments, declared to forced named binding of later arguments}

\item{minFraction}{optional minimum frequency a categorical level must have to be converted to an indicator column.}

\item{weights}{optional training weights for each row}

\item{rareCount}{optional integer, suppress direct effects of level of this count or less.}

\item{collarProb}{what fraction of the data (pseudo-probability) to collar data at (<0.5).}

\item{verbose}{if TRUE print progress.}

\item{parallelCluster}{(optional) a cluster object created by package parallel or package snow}
}
\value{
treatment plan (for use with prepare)
}
\description{
Data frame is assumed to have only atomic columns
except for dates (which are converted to numeric).
Note: each column is processed independently of all others.
}
\details{
The main fields are mostly vectors with names (all with the same names in the same order):

- vars : (character array without names) names of variables (in same order as names on the other diagnostic vectors)
- varMoves : logical TRUE if the variable varied during hold out scoring, only variables that move will be in the treated frame

See the vtreat vignette for a bit more detail and a worked example.
}
\examples{

dTrainZ <- data.frame(x=c('a','a','a','a','b','b',NA,'e','e'),
    z=c(1,2,3,4,5,6,7,NA,9))
dTestZ <- data.frame(x=c('a','x','c',NA),
    z=c(10,20,30,NA))
treatmentsZ = designTreatmentsZ(dTrainZ,colnames(dTrainZ),
  rareCount=0)
dTrainZTreated <- prepare(treatmentsZ,dTrainZ,pruneSig=1)
dTestZTreated <- prepare(treatmentsZ,dTestZ,pruneSig=1)

}
\seealso{
\code{\link{prepare}} \code{\link{designTreatmentsC}} \code{\link{designTreatmentsN}}
}

