% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{getNewVarNames}
\alias{getNewVarNames}
\title{Return a list of new treated variable names (coresponding to optional original variable names)}
\usage{
getNewVarNames(treatments, origVarNames = c())
}
\arguments{
\item{origVarNames}{optional restrict to only derived variable originating from these original variables (null is no restriction)}

\item{treatemnts}{the treatments slot from a treatmentplan object}
}
\value{
list of new treated variable names
}
\description{
Return a list of new treated variable names (coresponding to optional original variable names)
}

