% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{hold1OutMeans}
\alias{hold1OutMeans}
\title{of all but the i-th y.  Useful for normalizing PRESS style statistics.}
\usage{
hold1OutMeans(y, weights)
}
\arguments{
\item{y}{values to average (should not have NAs).}

\item{weights}{data weighing (should not have NAs, be non-negative and not all zero).}
}
\value{
a vector of length(y) where the i-th entry is the weighted mean
}
\description{
of all but the i-th y.  Useful for normalizing PRESS style statistics.
}
\seealso{
\code{\link{pressStatOfBestLinearFit}}
}

