% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{kWayCrossValidation}
\alias{kWayCrossValidation}
\title{k-fold cross validation, a splitFunction in the sense of vtreat::buildEvalSets}
\usage{
kWayCrossValidation(nRows, nSplits, dframe, y)
}
\arguments{
\item{nRows}{number of rows to split.}

\item{nSplits}{number of groups to split into.}

\item{dframe}{original data frame (ignored).}

\item{y}{numeric outcome variable (ignored).}
}
\value{
split plan
}
\description{
k-fold cross validation, a splitFunction in the sense of vtreat::buildEvalSets
}
\examples{

kWayCrossValidation(3,5,NULL,NULL)

}

