% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{oneWayHoldout}
\alias{oneWayHoldout}
\title{One way holdout, a splitFunction in the sense of vtreat::buildEvalSets}
\usage{
oneWayHoldout(nSplits, nRows, dframe, y)
}
\arguments{
\item{nSplits}{number of groups to split into.}

\item{nRows}{number of rows to split.}

\item{dframe}{original data frame.}

\item{y}{numeric outcome variable.}
}
\value{
split plan
}
\description{
One way holdout, a splitFunction in the sense of vtreat::buildEvalSets
}

