% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{pressStatOfBestLinearFit}
\alias{pressStatOfBestLinearFit}
\title{Compute the PRESS R-squared statistic of a 1-variable linear model}
\usage{
pressStatOfBestLinearFit(x, y, weights = c(), normalizationStrat = "total")
}
\arguments{
\item{x}{numeric (no NAs/NULLs) effective variable}

\item{y}{numeric (no NAs/NULLs) outcome variable}

\item{weights}{(optional) numeric, non-negative, no NAs/NULLs at least two positive positions}

\item{normalizationStrat}{(optional) 'none': no normalization (traditional PRESS), 'total': divide by total variation, 'holdout': divide by 1-hold out variation (PRESS-line, larger than total variation)}
}
\value{
PRESS statistic of model y ~ a*x + b divided by pressStatOfBestConstant(y,weights), it is an R-squared so 1 is good 0 is bad
}
\description{
Compute the PRESS R-squared statistic of a 1-variable linear model
}
\seealso{
\code{\link{hold1OutMeans}}
}

