% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{print.vtreatment}
\alias{print.vtreatment}
\title{Print treatmentplan.}
\usage{
\method{print}{vtreatment}(x, ...)
}
\arguments{
\item{x}{treatmentplan}

\item{...}{additional args (to match general signature).}
}
\description{
Print treatmentplan.
}
\seealso{
\code{\link{designTreatmentsC}} \code{\link{designTreatmentsN}} \code{\link{prepare}}
}

