% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{problemAppPlan}
\alias{problemAppPlan}
\title{check if appPlan is a good carve-up of 1:nRows into nSplits groups}
\usage{
problemAppPlan(nRows, nSplits, appPlan, strictCheck)
}
\arguments{
\item{nRows}{number of rows to carve-up}

\item{nSplits}{number of sets to carve-up into}

\item{appPlan}{carve-up to critique}

\item{strictCheck}{logical, if true expect application data to be a carve-up and training data to be a maximal partition and to match nSplits.}
}
\value{
problem with carve-up (null if good)
}
\description{
check if appPlan is a good carve-up of 1:nRows into nSplits groups
}
\examples{

plan <- kWayStratifiedY(3,2,NULL,NULL)
problemAppPlan(3,3,plan,TRUE)

}
\seealso{
\code{\link{kWayCrossValidation}}, \code{\link{kWayStratifiedY}}, and \code{\link{makekWayCrossValidationGroupedByColumn}}
}
