% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfSample.R
\name{problemAppPlan}
\alias{problemAppPlan}
\title{check if appPlan is a good partition of 1:nRows into nSplits groups}
\usage{
problemAppPlan(nRows, nSplits, appPlan)
}
\arguments{
\item{nRows}{number of rows to partition}

\item{nSplits}{number of sets to partition into}

\item{appPlan}{partition to critique}
}
\value{
problem with partition (null if good)
}
\description{
check if appPlan is a good partition of 1:nRows into nSplits groups
}

