% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapr_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{apply_left}
\alias{apply_right}
\alias{mk_tmp_name_source}
\alias{map_to_char}
\alias{\%.>\%}
\alias{let}
\alias{qc}
\alias{qe}
\alias{qae}
\alias{build_frame}
\alias{draw_frame}
\alias{qchar_frame}
\alias{\%:=\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{wrapr}{\code{\link[wrapr]{\%:=\%}}, \code{\link[wrapr]{\%.>\%}}, \code{\link[wrapr]{apply_left}}, \code{\link[wrapr]{apply_right}}, \code{\link[wrapr]{build_frame}}, \code{\link[wrapr]{draw_frame}}, \code{\link[wrapr]{let}}, \code{\link[wrapr]{map_to_char}}, \code{\link[wrapr]{mk_tmp_name_source}}, \code{\link[wrapr]{qae}}, \code{\link[wrapr]{qc}}, \code{\link[wrapr]{qchar_frame}}, \code{\link[wrapr]{qe}}}
}}

