% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{vnames}
\alias{vnames}
\title{New treated variable names from a treatmentplan$treatment item.}
\usage{
vnames(x)
}
\arguments{
\item{x}{vtreatment item}
}
\description{
New treated variable names from a treatmentplan$treatment item.
}
\seealso{
\code{\link{designTreatmentsC}} \code{\link{designTreatmentsN}}
}

