% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vtreat.R
\name{vorig}
\alias{vorig}
\title{Original variable name.}
\usage{
vorig(x, ...)
}
\arguments{
\item{x}{vtreatment item.}

\item{...}{additional args (to match general signature)}
}
\description{
Original variable name.
}

